# Refactor: performance optimization in tracker announce