--- a/packages/torrent-repository-benchmarks/src/benches/asyn.rs
+++ b/packages/torrent-repository-benchmarks/src/benches/asyn.rs
@@ -19,9 +19,7 @@ pub async fn async_add_one_torrent<T: TRepositoryAsync + Send + Sync + 'static>(
 
         let start_time = std::time::Instant::now();
 
-        torrent_repository
-            .update_torrent_with_peer_and_get_stats(&info_hash, &DEFAULT_PEER)
-            .await;
+        torrent_repository.announce(&info_hash, &DEFAULT_PEER).await;
 
         let result = start_time.elapsed();
 
@@ -45,9 +43,7 @@ pub async fn async_update_one_torrent_in_parallel<T: TRepositoryAsync + Send + S
         let handles = FuturesUnordered::new();
 
         // Add the torrent/peer to the torrent repository
-        torrent_repository
-            .update_torrent_with_peer_and_get_stats(info_hash, &DEFAULT_PEER)
-            .await;
+        torrent_repository.announce(info_hash, &DEFAULT_PEER).await;
 
         let start_time = std::time::Instant::now();
 
@@ -55,9 +51,7 @@ pub async fn async_update_one_torrent_in_parallel<T: TRepositoryAsync + Send + S
             let torrent_repository_clone = torrent_repository.clone();
 
             let handle = runtime.spawn(async move {
-                torrent_repository_clone
-                    .update_torrent_with_peer_and_get_stats(info_hash, &DEFAULT_PEER)
-                    .await;
+                torrent_repository_clone.announce(info_hash, &DEFAULT_PEER).await;
 
                 if let Some(sleep_time) = args.sleep {
                     let start_time = std::time::Instant::now();
@@ -99,9 +93,7 @@ pub async fn async_add_multiple_torrents_in_parallel<T: TRepositoryAsync + Send 
             let torrent_repository_clone = torrent_repository.clone();
 
             let handle = runtime.spawn(async move {
-                torrent_repository_clone
-                    .update_torrent_with_peer_and_get_stats(&info_hash, &DEFAULT_PEER)
-                    .await;
+                torrent_repository_clone.announce(&info_hash, &DEFAULT_PEER).await;
 
                 if let Some(sleep_time) = args.sleep {
                     let start_time = std::time::Instant::now();
@@ -139,9 +131,7 @@ pub async fn async_update_multiple_torrents_in_parallel<T: TRepositoryAsync + Se
 
         // Add the torrents/peers to the torrent repository
         for info_hash in &info_hashes {
-            torrent_repository
-                .update_torrent_with_peer_and_get_stats(info_hash, &DEFAULT_PEER)
-                .await;
+            torrent_repository.announce(info_hash, &DEFAULT_PEER).await;
         }
 
         let start_time = std::time::Instant::now();
@@ -150,9 +140,7 @@ pub async fn async_update_multiple_torrents_in_parallel<T: TRepositoryAsync + Se
             let torrent_repository_clone = torrent_repository.clone();
 
             let handle = runtime.spawn(async move {
-                torrent_repository_clone
-                    .update_torrent_with_peer_and_get_stats(&info_hash, &DEFAULT_PEER)
-                    .await;
+                torrent_repository_clone.announce(&info_hash, &DEFAULT_PEER).await;
 
                 if let Some(sleep_time) = args.sleep {
                     let start_time = std::time::Instant::now();