--- a/tests/servers/http/environment.rs
+++ b/tests/servers/http/environment.rs
@@ -20,7 +20,7 @@ pub struct Environment<S> {
 impl<S> Environment<S> {
     /// Add a torrent to the tracker
     pub async fn add_torrent_peer(&self, info_hash: &InfoHash, peer: &Peer) {
-        self.tracker.update_torrent_with_peer_and_get_stats(info_hash, peer).await;
+        self.tracker.inner_announce(info_hash, peer).await;
     }
 }