@da2ce7 2024-03-22
I would prefer to merge #690 before this. I've just got to finish a couple more tests and then it is ready to review.  (I will prob need to do rebase to squash a few comments also.)