@josecelano 2024-03-22
> I would prefer to merge #690 before this. I've just got to finish a couple more tests and then it is ready to review.  (I will prob need to do rebase to squash a few comments also.)

Hey, yes I didn't want to merge it before you finish yours. Anyway It's not an important change. Take a look when you have time and let me know if you think the code is simpler. Regarding performance It looks like it's not a big change.