@josecelano 2024-04-12
I'm closing this. I think it's a good refactor. I would like to decouple peer lists from statistics. Statistics could even be enabled/disabled in config. Since the performance is the same, I would prefer to split the method `update_torrent_with_peer_and_get_stats` ([command/query segregation](https://en.wikipedia.org/wiki/Command_Query_Responsibility_Segregation)).

In short, I think it would be better because:

- Stats could be optional.
- Methos are simpler command/query segregation.